# docs/api/openapi-v0.1.yaml
openapi: 3.1.0
info:
  title: Tascade API
  version: 0.1.0
  description: |
    Canonical REST/JSON API for the Tascade coordinator.

    Design constraints:
    - Single source of truth for task/dependency/gate invariants.
    - MCP adapter is a thin proxy over these endpoints.
    - Idempotency required for mutating operations.
    - Project-scoped API keys required for all operations.
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
...
paths:
  /v1/projects:
    post:
      tags: [Projects]
      summary: Create project
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Error'
  /v1/tasks:
    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        default:
          $ref: '#/components/responses/Error'
  /v1/dependencies:
    post:
      tags: [Dependencies]
      summary: Create dependency edge
      operationId: createDependency
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDependencyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyEdge'
        '409':
          description: Conflict (e.g., cycle detected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/tasks/{taskId}/claim:
    post:
      tags: [Tasks]
      summary: Claim a task and acquire lease
      operationId: claimTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimTaskRequest'
      responses:
        '200':
          description: Claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimTaskResponse'
        '409':
          description: Not claimable (lease/reservation/state conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/tasks/{taskId}/assign:
    post:
      tags: [Tasks]
      summary: Create hard reservation for a specific assignee
      operationId: assignTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTaskRequest'
      responses:
        '200':
          description: Reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskReservation'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/tasks/{taskId}/heartbeat:
    post:
      tags: [Tasks]
      summary: Extend lease and report seen plan version
      operationId: heartbeatTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '200':
          description: Heartbeat accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
        '409':
          description: Plan stale or invalid lease token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/plans/changesets:
    post:
      tags: [Planning]
      summary: Create plan change set
      operationId: createPlanChangeset
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanChangesetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanChangeset'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/plans/changesets/{changesetId}/apply:
    post:
      tags: [Planning]
      summary: Apply plan change set atomically
      operationId: applyPlanChangeset
      parameters:
        - $ref: '#/components/parameters/ChangesetIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyPlanChangesetRequest'
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyPlanChangesetResponse'
        '409':
          description: Stale base plan version or invariant violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional client-provided idempotency key for safe retries.
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    TaskIdPath:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ChangesetIdPath:
      name: changesetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Machine error code, e.g. CYCLE_DETECTED, PLAN_STALE,
            LEASE_INVALID, RESERVATION_CONFLICT, INVARIANT_VIOLATION.
        message:
          type: string
        retryable:
          type: boolean
        correlation_id:
          type: string
        details:
          type: object
          additionalProperties: true
    Project:
      type: object
      required: [id, name, status, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        org_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        status:
          type: string
          enum: [active, paused, archived]
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    WorkSpec:
      type: object
      description: Structured execution spec consumed by agents.
      required: [objective, acceptance_criteria]
      properties:
        objective:
          type: string
        constraints:
          type: array
          items:
            type: string
        acceptance_criteria:
          type: array
          items:
            type: string
        interfaces:
          type: array
          items:
            type: string
        path_hints:
          type: array
          items:
            type: string
    Task:
      type: object
      required:
        - id
        - project_id
        - title
        - state
        - priority
        - work_spec
        - task_class
        - capability_tags
        - version
        - created_at
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        phase_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          nullable: true
        milestone_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        state:
          $ref: '#/components/schemas/TaskState'
        priority:
          type: integer
        work_spec:
          $ref: '#/components/schemas/WorkSpec'
        task_class:
          $ref: '#/components/schemas/TaskClass'
        capability_tags:
          type: array
          items:
            type: string
        expected_touches:
          type: array
          items:
            type: string
        exclusive_paths:
          type: array
          items:
            type: string
        shared_paths:
          type: array
          items:
            type: string
        introduced_in_plan_version:
          type: integer
          nullable: true
        deprecated_in_plan_version:
          type: integer
          nullable: true
        version:
          type: integer
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    TaskState:
      type: string
      enum:
        - backlog
        - ready
        - reserved
        - claimed
        - in_progress
        - implemented
        - integrated
        - conflict
        - blocked
        - abandoned
        - cancelled
    TaskClass:
      type: string
      enum:
        - architecture
        - db_schema
        - security
        - cross_cutting
        - review_gate
        - merge_gate
        - frontend
        - backend
        - crud
        - other
    DependencyEdge:
      type: object
      required: [id, project_id, from_task_id, to_task_id, unlock_on]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        from_task_id:
          $ref: '#/components/schemas/UUID'
        to_task_id:
          $ref: '#/components/schemas/UUID'
        unlock_on:
          type: string
          enum: [implemented, integrated]
    Lease:
      type: object
      required: [id, task_id, agent_id, token, status, expires_at, heartbeat_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        task_id:
          $ref: '#/components/schemas/UUID'
        agent_id:
          type: string
        token:
          type: string
        status:
          type: string
          enum: [active, expired, released, consumed]
        expires_at:
          $ref: '#/components/schemas/Timestamp'
        heartbeat_at:
          $ref: '#/components/schemas/Timestamp'
    TaskReservation:
      type: object
      required: [id, task_id, assignee_agent_id, mode, status, ttl_seconds, expires_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        task_id:
          $ref: '#/components/schemas/UUID'
        assignee_agent_id:
          type: string
        mode:
          type: string
          enum: [hard]
        status:
          type: string
          enum: [active, expired, released, consumed]
        ttl_seconds:
          type: integer
        expires_at:
          $ref: '#/components/schemas/Timestamp'
        created_at:
          $ref: '#/components/schemas/Timestamp'
...
    PlanVersion:
      type: object
      required: [id, project_id, version_number, created_at, created_by]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        version_number:
          type: integer
        change_set_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          nullable: true
        summary:
          type: string
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        created_by:
          type: string
    PlanChangeset:
      type: object
      required: [id, project_id, base_plan_version, target_plan_version, status, operations, created_at, created_by]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        base_plan_version:
          type: integer
        target_plan_version:
          type: integer
        status:
          type: string
          enum: [draft, validated, applied, rejected]
        operations:
          type: array
          items:
            $ref: '#/components/schemas/PlanOperation'
        impact_preview:
          type: object
          additionalProperties: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        created_by:
          type: string
        applied_at:
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
    PlanOperation:
      type: object
      required: [op]
      properties:
        op:
          type: string
          enum:
            - add_task
            - update_task
            - deprecate_task
            - reprioritize_task
            - add_dependency
            - remove_dependency
            - move_phase
            - update_unlock_on
        task_id:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
    TaskExecutionSnapshot:
      type: object
      required: [id, task_id, lease_id, captured_plan_version, work_spec_hash, work_spec_payload, captured_at, captured_by]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        task_id:
          $ref: '#/components/schemas/UUID'
        lease_id:
          $ref: '#/components/schemas/UUID'
        captured_plan_version:
          type: integer
        work_spec_hash:
          type: string
        work_spec_payload:
          $ref: '#/components/schemas/WorkSpec'
        captured_at:
          $ref: '#/components/schemas/Timestamp'
        captured_by:
          type: string
...
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        org_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
    CreateTaskRequest:
      type: object
      required: [project_id, title, work_spec, task_class]
      properties:
        project_id:
          $ref: '#/components/schemas/UUID'
        phase_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          nullable: true
        milestone_id:
          allOf:
            - $ref: '#/components/schemas/UUID'
          nullable: true
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: integer
          default: 100
        work_spec:
          $ref: '#/components/schemas/WorkSpec'
        task_class:
          $ref: '#/components/schemas/TaskClass'
        capability_tags:
          type: array
          items:
            type: string
        expected_touches:
          type: array
          items:
            type: string
        exclusive_paths:
          type: array
          items:
            type: string
        shared_paths:
          type: array
          items:
            type: string
    CreateDependencyRequest:
      type: object
      required: [project_id, from_task_id, to_task_id, unlock_on]
      properties:
        project_id:
          $ref: '#/components/schemas/UUID'
        from_task_id:
          $ref: '#/components/schemas/UUID'
        to_task_id:
          $ref: '#/components/schemas/UUID'
        unlock_on:
          type: string
          enum: [implemented, integrated]
...
    ClaimTaskRequest:
      type: object
      required: [project_id, agent_id]
      properties:
        project_id:
          $ref: '#/components/schemas/UUID'
        agent_id:
          type: string
        claim_mode:
          type: string
          enum: [pull, directed]
          default: pull
        seen_plan_version:
          type: integer
          nullable: true
    ClaimTaskResponse:
      type: object
      required: [task, lease, execution_snapshot]
      properties:
        task:
          $ref: '#/components/schemas/Task'
        lease:
          $ref: '#/components/schemas/Lease'
        execution_snapshot:
          $ref: '#/components/schemas/TaskExecutionSnapshot'
    AssignTaskRequest:
      type: object
      required: [project_id, assignee_agent_id, created_by]
      properties:
        project_id:
          $ref: '#/components/schemas/UUID'
        assignee_agent_id:
          type: string
        ttl_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          default: 1800
        created_by:
          type: string
...
    HeartbeatRequest:
      type: object
      required: [project_id, agent_id, lease_token]
      properties:
        project_id:
          $ref: '#/components/schemas/UUID'
        agent_id:
          type: string
        lease_token:
          type: string
        seen_plan_version:
          type: integer
          nullable: true
    HeartbeatResponse:
      type: object
      required: [lease_expires_at, plan_version]
      properties:
        lease_expires_at:
          $ref: '#/components/schemas/Timestamp'
        plan_version:
          type: integer
        stale:
          type: boolean
          default: false
        stale_action:
          type: string
          enum: [refresh, continue_with_notice, human_review]
          nullable: true
...
    CreatePlanChangesetRequest:
      type: object
      required: [project_id, base_plan_version, target_plan_version, operations, created_by]
      properties:
        project_id:
          $ref: '#/components/schemas/UUID'
        base_plan_version:
          type: integer
        target_plan_version:
          type: integer
        operations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PlanOperation'
        created_by:
          type: string
    ApplyPlanChangesetRequest:
      type: object
      properties:
        allow_rebase:
          type: boolean
          default: false
        applied_by:
          type: string
    ApplyPlanChangesetResponse:
      type: object
      required: [changeset, plan_version]
      properties:
        changeset:
          $ref: '#/components/schemas/PlanChangeset'
        plan_version:
          $ref: '#/components/schemas/PlanVersion'
        invalidated_claim_task_ids:
          type: array
          items:
            $ref: '#/components/schemas/UUID'
        invalidated_reservation_task_ids:
          type: array
          items:
            $ref: '#/components/schemas/UUID'
...